Move RepoUniverseSymbolSearchJob into zoekt package Since this job runs a Zoekt search, it makes more sense for it to live in the zoekt package alongside ZoektSymbolSearchJob. This commit moves the job and updates references.
